Fix dangling ../shared imports after shared->core rename
