+++ packages/upscalerjs/src/node/image.node.ts
@@ -1,9 +1,9 @@
 import fs from 'fs';
 import type { Tensor, Tensor3D, Tensor4D, } from '@tensorflow/tfjs-core';
 import type { TFN, } from '../../../shared/src/types';
-import { tensorAsClampedArray, } from '../shared/tensor-utils';
+import { tensorAsClampedArray, } from '../core/tensor-utils';
 import { isFourDimensionalTensor, isThreeDimensionalTensor, isTensor, isString, hasValidChannels, } from '../../../shared/src/constants';
-import { CheckValidEnvironment, GetImageAsTensor, TensorAsBase64, } from '../shared/types';
+import { CheckValidEnvironment, GetImageAsTensor, TensorAsBase64, } from '../core/types';
 
 export const getInvalidTensorError = (input: Tensor): Error => new Error(
   [
